COCO-Counterfactuals/coco-counterfactuals-src/requirements.txt
tqdm
nltk
torch>=1.5.0
torchvision>=0.6.0
datasets>=1.8.0
sentence_transformers
pyarrow
pandas
